Extract chat rendering helpers
