+++ tests/test-correctness-phase2.js
@@ -132,6 +132,7 @@ const pwaAppShellAssets = [
   '/js/chat-panel.js',
   '/js/chat-discussion.js',
   '/js/chat-onboarding.js',
+  '/js/chat-render.js',
   '/js/chat-icons.js',
   '/js/chat-personalities.js',
   '/js/chat-history.js',
